requirements.txt
# Core framework
PyQt5==5.15.11
SQLAlchemy==2.0.36
alembic==1.14.1

# Camera and image processing
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.4.0

# Face detection/recognition
face-recognition==1.3.0
dlib==19.24.6

# AI hairstyle generation
torch==2.5.1
diffusers==0.31.0
transformers==4.47.0
accelerate==1.2.1
safetensors==0.4.5

# Utilities
python-dotenv==1.0.1
msgpack==1.1.0
